# api.yaml
openapi: 3.0.3
info:
  title: Dayflow HRMS API
  version: 1.1.0
  description: Canonical frontend/backend API contract aligned with the backend plan and project documentation.
...
paths:
...
  /simulation/history/{user_id}:
    get:
      tags: [Simulation]
      summary: Get simulation history for a user
      x-access: Admin/HR
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Simulation history returned
...
components:
  parameters:
...
    UserId:
      in: path
      name: user_id
      required: true
      schema:
        type: integer